+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Refactor post reactions [#325](https://github.com/rokwire/groups-building-block/issues/325)
 
 ## [1.15.0] - 2023-02-01
 ### Added